scm: move dulwich into scm.backend
